requirements.txt
pandas
matplotlib
datasets
huggingface_hub
Pillow
